unified ms/engine test file (#928)
